Replace Login with CatalogModel
