+++ provider_shopper/lib/models/cart.dart
@@ -1,24 +1,7 @@
 import 'dart:collection';
 
 import 'package:flutter/foundation.dart';
-
-const _itemNames = [
-  'Code Smell',
-  'Control Flow',
-  'Interpreter',
-  'Recursion',
-  'Sprint',
-  'Heisenbug',
-  'Spaghetti',
-  'Hydra Code',
-  'Off-By-One',
-  'Scope',
-  'Callback',
-  'Closure',
-  'Automata',
-  'Bit Shift',
-  'Currying',
-];
+import 'package:provider_shopper/models/catalog.dart';
 
 class CartModel extends ChangeNotifier {
   /// Internal, private state of the cart.
@@ -37,18 +20,4 @@ class CartModel extends ChangeNotifier {
     // depend on it.
     notifyListeners();
   }
-}
-
-@immutable
-class Item {
-  final int id;
-  final String name;
-
-  Item(this.id) : name = _itemNames[id % _itemNames.length];
-
-  @override
-  int get hashCode => id;
-
-  @override
-  bool operator ==(other) => other is Item && other.id == id;
 }
