+++ provider_shopper/lib/models/catalog.dart
@@ -1,0 +1,49 @@
+import 'package:flutter/material.dart';
+
+/// A proxy of the catalog of items the user can buy.
+///
+/// In a real app, this might be backed by a backend and cached on device.
+/// In this sample app, the catalog is procedurally generated.
+///
+/// For simplicity, the catalog is expected to be immutable (no products are
+/// expected to be added, removed or changed during the execution of the app).
+class CatalogModel {
+  static const _itemNames = [
+    'Code Smell',
+    'Control Flow',
+    'Interpreter',
+    'Recursion',
+    'Sprint',
+    'Heisenbug',
+    'Spaghetti',
+    'Hydra Code',
+    'Off-By-One',
+    'Scope',
+    'Callback',
+    'Closure',
+    'Automata',
+    'Bit Shift',
+    'Currying',
+  ];
+
+  /// Get item by its position in the catalog.
+  Item getByPosition(int index) {
+    return Item(index, _itemNames[index % _itemNames.length]);
+  }
+}
+
+@immutable
+class Item {
+  final int id;
+  final String name;
+  final Color color;
+
+  Item(this.id, this.name)
+      : color = Colors.primaries[id % Colors.primaries.length];
+
+  @override
+  int get hashCode => id;
+
+  @override
+  bool operator ==(other) => other is Item && other.id == id;
+}
